src/main/java/mx/iteso/decorator/condiments/TortillaHarina.java
package mx.iteso.decorator.condiments;

import mx.iteso.decorator.CondimentsDecorator;
import mx.iteso.decorator.Taco;

public class TortillaHarina extends CondimentsDecorator {
    Taco taco;

    public TortillaHarina(Taco taco){
        this.taco = taco;
        this.size = taco.size;
    }

    @Override
    public String getDescription() {
        return taco.getDescription() + " en tortilla de harina,";
    }

    @Override
    public double cost() throws Exception {
        return 4 + taco.cost();
    }
}
